deletes disclaimer dir, moves utils to common/utils/provenance
