+++ src/components/LocationPage/ChartsHolder.style.tsx
@@ -2,6 +2,7 @@ import styled, { css } from 'styled-components';
 import palette from 'assets/theme/palette';
 import { Typography } from '@material-ui/core';
 import { mobileBreakpoint, materialSMBreakpoint } from 'assets/theme/sizes';
+import { COLOR_MAP } from 'common/colors';
 
 export const ChartContentWrapper = styled.div`
   margin-top: 85px;
@@ -106,4 +107,15 @@ export const BetaTag = styled.div`
   background-color: ${palette.info.main};
   color: white;
   transform: translateY(-0.375rem);
+`;
+
+export const DisclaimerWrapper = styled.div`
+  max-width: 600px;
+  padding: 1.5rem 0 0.75rem;
+  color: ${COLOR_MAP.GRAY_BODY_COPY};
+  font-size: 0.875rem;
+
+  @media (min-width: ${materialSMBreakpoint}) {
+    padding: 1.5rem 0 2rem;
+  }
 `;
